Move controller tests out of `presroc` folder
